+++ Server/models/upgrade.js
@@ -18,7 +18,7 @@ module.exports = (sequelize, DataTypes) => {
     }
   }
   Upgrade.init({
-    upgrade: DataTypes.INTEGER,
+    upgrade: DataTypes.FLOAT,
     pkg_price: DataTypes.INTEGER,
     level: DataTypes.INTEGER,
     user_id: DataTypes.INTEGER,
